Clean up connector integration tests
