+++ xyz-psql-connector/src/test/resources/events/HealthCheckEvent.json
@@ -3,7 +3,6 @@
 	"minResponseTime": 100,
 	"connectorParams": {
 		"propertySearch": true,
-		"autoIndexing" : true,
 		"connectorId": "test-connector"
 	},
 	"params": {
